fix: resolve warnings from tests (#509)
* resolved warnings: incompatible numpy dtype, unary_union -> union_all(), ensure same crs, iterating over datacatalog

* investigated all warnings: fixed some, ignored some, and error + TODO some

* fix to float32

* resolve Geometry is in a geographic CRS. Results from <planar_operation> are likely incorrect :UserWarning

* dont error on the warnings: `driver .* does not support open option` and `Dataset has no geotransform, gcps, or rpcs.`

* cleanup

* update lockfile

* small updates

+++ hydromt_wflow/workflows/glaciers.py
@@ -115,8 +115,12 @@ storage per grid cell"
                 3857
             ).area  # area calculation needs projected crs
             gfrac = garea_cell / np.sum(garea_cell)
-            gdf_grid.loc[idxs, "glacierfrac"] += garea_cell / gdf_grid.loc[idxs, "area"]
-            gdf_grid.loc[idxs, "glacierstore"] += gfrac * glacier["glacierstore"]
+            gdf_grid.loc[idxs, "glacierfrac"] += (
+                garea_cell / gdf_grid.loc[idxs, "area"]
+            ).astype(np.float32)
+            gdf_grid.loc[idxs, "glacierstore"] += (
+                gfrac * glacier["glacierstore"]
+            ).astype(np.float32)
 
     # reproject back to original projection
     # Create the rasterized glacier storage map
